Updates based on review comments.
Renamed eventflow packages, added copyrights, removed unused code and comments
in java integration test operators.
Rationalize used versions (stub vs integration test) in POMs.
Fixed version in documentation.
