--- ep-maven/src/it/EventFlowPlusApp/SimpleEventFlow/src/main/eventflow/com/tibco/perf/myeventflow/MyEventFlow.sbapp
@@ -1,36 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="no"?>
-<modify version="11.0.0-SNAPSHOT_f3979de6c042595714bb2e097efff505b9f2ed17">
-    <add>
-        <annotations>
-            <annotation name="hygienic"/>
-        </annotations>
-        <type-metadata>
-            <param name="type" value="module"/>
-            <param name="fully-qualified-name" value="com.tibco.perf.myeventflow.MyEventFlow"/>
-        </type-metadata>
-        <memory-model-settings/>
-        <dynamic-variables/>
-        <stream name="InputStream">
-            <schema>
-                <field name="name" type="string"/>
-            </schema>
-        </stream>
-        <box name="Map" type="map">
-            <input port="1" stream="InputStream"/>
-            <output port="1" stream="out:Map_1"/>
-            <target-list>
-                <item name="input" selection="all"/>
-                <expressions>
-                    <include field="other">"abc"</include>
-                </expressions>
-            </target-list>
-        </box>
-        <box name="CustomOperator" type="java">
-            <input port="1" stream="out:Map_1"/>
-            <output port="1" stream="OutputStream"/>
-            <param name="start:state" value="true"/>
-            <param name="javaclass" value="com.tibco.myeventflow.MyOperator"/>
-        </box>
-        <output-stream name="OutputStream"/>
-    </add>
-</modify>
