--- ep-maven/src/it/EventFlowPlusApp/SimpleEventFlow/src/main/eventflow/com/tibco/perf/myeventflow/MyEventFlow.sblayout
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?><application-layout>
-    <component name="InputStream">
-        <param name="gui:x" value="-347"/>
-        <param name="gui:y" value="-91"/>
-    </component>
-    <component name="Map">
-        <param name="gui:x" value="-260"/>
-        <param name="gui:y" value="-91"/>
-    </component>
-    <component name="CustomOperator">
-        <param name="gui:x" value="-163"/>
-        <param name="gui:y" value="-91"/>
-    </component>
-    <component name="OutputStream">
-        <param name="gui:x" value="-46"/>
-        <param name="gui:y" value="-91"/>
-    </component>
-</application-layout>
