--- ep-maven/src/it/EventFlowPlusApp/SimpleEventFlow/src/main/java/com/tibco/myeventflow/MyOperatorBeanInfo.java
@@ -1,22 +1,0 @@
-package com.tibco.myeventflow;
-
-import com.streambase.sb.operator.parameter.*;
-
-/**
-* A BeanInfo class controls what properties are exposed, add 
-* metadata about properties (such as which properties are optional), and access 
-* special types of properties that can't be automatically derived via reflection. 
-* If a BeanInfo class is present, only the properties explicitly declared in
-* this class will be exposed by StreamBase.
-*/
-public class MyOperatorBeanInfo extends SBSimpleBeanInfo {
-
-	/*
-	* The order of properties below determines the order they are displayed within
-	* the StreamBase Studio property view. 
-	*/
-	public SBPropertyDescriptor[] getPropertyDescriptorsChecked() {
-		return new SBPropertyDescriptor[] {};
-	}
-
-}
