+++ ep-maven/src/it/EventFlowPlusApp/invoker.properties
@@ -29,4 +29,4 @@
 #
 
 
-invoker.goals = install integration-test
+invoker.goals = install
