+++ ep-maven/src/it/SimpleEventFlow/src/test/java/com/tibco/test/myeventflow/MyEventFlowTest.java
@@ -64,13 +64,10 @@ public class MyEventFlowTest extends UnitTest {
 	 */
 	@BeforeClass
 	public static void setupServer() throws StreamBaseException, ConfigurationException, InterruptedException {
-		// Example configuration load
-		// Configuration.forFile("engine.conf").load().activate();
 
-		// create a StreamBase server and load modules once for all tests in this class
 		server = ServerManagerFactory.getEmbeddedServer();
 		server.startServer();
-		server.loadApp("com.tibco.perf.myeventflow.MyEventFlow");
+		server.loadApp("com.tibco.test.MyEventFlow");
 	}
 
 	/**
