+++ ep-maven/src/site/markdown/liveview_examples.md
@@ -49,7 +49,7 @@ repository, a LiveView fragment.
 When the maven install goal is called (mvn install), this pom.xml instructs
 maven to perform the following steps :
   
-1. Uses [install-product](https://tibcosoftware.github.io/tibco-streaming-maven-plugin/1.7.0-SNAPSHOT/ep-maven-plugin/install-product-mojo.html) to check if the 
+1. Uses [install-product](https://tibcosoftware.github.io/tibco-streaming-maven-plugin/2.0.0-SNAPSHOT/ep-maven-plugin/install-product-mojo.html) to check if the 
     dependent product ( in this com.tibco.ep.thirdparty:tibco-sb_linuxx86_64 ) is
     installed.  If its not, maven will download the archive and the plugin
     will extract into $TIBCO_EP_HOME.
@@ -57,19 +57,19 @@ maven to perform the following steps :
 2. Uses the standard maven plugin [maven-compiler-plugin:testCompile](https://maven.apache.org/plugins/maven-compiler-plugin/testCompile-mojo.html)
     to compile any java test sources to class files.
     
-3. Uses [start-nodes](https://tibcosoftware.github.io/tibco-streaming-maven-plugin/1.7.0-SNAPSHOT/ep-maven-plugin/start-nodes-mojo.html) to start a test cluster.  
+3. Uses [start-nodes](https://tibcosoftware.github.io/tibco-streaming-maven-plugin/2.0.0-SNAPSHOT/ep-maven-plugin/start-nodes-mojo.html) to start a test cluster.  
     Since this pom.xml has no configuration, a single node is started 
     A.$\{artifactId\} (ie A.goldylocks in this example) with a random but unused 
     discovery port.
     
-4. Uses [test-liveview-fragment](https://tibcosoftware.github.io/tibco-streaming-maven-plugin/1.7.0-SNAPSHOT/ep-maven-plugin/test-liveview-fragment-mojo.html) 
+4. Uses [test-liveview-fragment](https://tibcosoftware.github.io/tibco-streaming-maven-plugin/2.0.0-SNAPSHOT/ep-maven-plugin/test-liveview-fragment-mojo.html) 
     to launch sbunit on the cluster and report the test results.  Should the 
     test cases fail then no further processing occurs.
     
-5. Uses [stop-nodes](https://tibcosoftware.github.io/tibco-streaming-maven-plugin/1.7.0-SNAPSHOT/ep-maven-plugin/stop-nodes-mojo.html) to stop and remove the test 
+5. Uses [stop-nodes](https://tibcosoftware.github.io/tibco-streaming-maven-plugin/2.0.0-SNAPSHOT/ep-maven-plugin/stop-nodes-mojo.html) to stop and remove the test 
     nodes
   
-6. Uses [package-liveview-fragment](https://tibcosoftware.github.io/tibco-streaming-maven-plugin/1.7.0-SNAPSHOT/ep-maven-plugin/package-liveview-fragment-mojo.html) 
+6. Uses [package-liveview-fragment](https://tibcosoftware.github.io/tibco-streaming-maven-plugin/2.0.0-SNAPSHOT/ep-maven-plugin/package-liveview-fragment-mojo.html) 
     to create a LiveView fragment zip file in the build directory (by 
     default, set to target) and attaches it to the build.  Note that the fragment
     zip will include any LiveView Web plugins, themes and wars.
@@ -83,7 +83,7 @@ maven to perform the following steps :
 ## Launch liveview fragment
 
 Whilst launching of liveview is usually done when an application archive is
-deployed, its possible to start the fragment using [deploy-fragment](https://tibcosoftware.github.io/tibco-streaming-maven-plugin/1.7.0-SNAPSHOT/ep-maven-plugin/deploy-fragment-mojo.html).
+deployed, its possible to start the fragment using [deploy-fragment](https://tibcosoftware.github.io/tibco-streaming-maven-plugin/2.0.0-SNAPSHOT/ep-maven-plugin/deploy-fragment-mojo.html).
 For example a test case might want to start liveview and use playback to inject
 test data.
