+++ stub_platform/sb/pom-sb-linuxaarch64.xml
@@ -43,7 +43,7 @@
     <parent>
         <groupId>com.tibco.ep.ep-maven</groupId>
         <artifactId>stub-platform</artifactId>
-        <version>1.7.0-SNAPSHOT</version>
+        <version>2.0.0-SNAPSHOT</version>
         <relativePath>..</relativePath>
     </parent>
